+++ auth_oidc_test.go
@@ -273,6 +273,7 @@ func TestAuthFlowOIDC(t *testing.T) { //nolint:dupl
 
 	verifyAllActivities(t, ts, service, accessToken, []ActivityExpectation{
 		{charon.ActivitySignIn, nil, []charon.Provider{"oidcTesting"}, 0, 1, 0, 1},
+		{charon.ActivityOrganizationUpdate, []charon.ActivityChangeType{charon.ActivityChangePermissionsAdded}, nil, 1, 1, 0, 0},
 		{charon.ActivityIdentityUpdate, []charon.ActivityChangeType{charon.ActivityChangeMembershipAdded}, nil, 1, 1, 0, 1},
 		{charon.ActivityIdentityCreate, nil, nil, 1, 0, 0, 0},
 	})
@@ -286,6 +287,7 @@ func TestAuthFlowOIDC(t *testing.T) { //nolint:dupl
 		{charon.ActivitySignIn, nil, []charon.Provider{"oidcTesting"}, 0, 1, 0, 1}, // Signin.
 		{charon.ActivitySignOut, nil, nil, 0, 0, 0, 0},                             // Signout.
 		{charon.ActivitySignIn, nil, []charon.Provider{"oidcTesting"}, 0, 1, 0, 1},
+		{charon.ActivityOrganizationUpdate, []charon.ActivityChangeType{charon.ActivityChangePermissionsAdded}, nil, 1, 1, 0, 0},
 		{charon.ActivityIdentityUpdate, []charon.ActivityChangeType{charon.ActivityChangeMembershipAdded}, nil, 1, 1, 0, 1},
 		{charon.ActivityIdentityCreate, nil, nil, 1, 0, 0, 0},
 	})